pyproject.toml
[project]
name = "navigai-api"
version = "0.1.0"
description = "AI-powered career navigation and mock interview platform"
authors = [
    {name = "Shashidhar Kundur",email = "shashidharkundur1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "firebase-admin (>=7.1.0,<8.0.0)",
    "pydantic (>=2.11.7,<3.0.0)",
    "quart (>=0.20.0,<0.21.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "quart-cors (>=0.8.0,<0.9.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "instructor (>=1.11.2,<2.0.0)",
    "generativeai (>=0.0.1,<0.0.2)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "openai-whisper @ git+https://github.com/openai/whisper.git",
    "pyttsx3 (>=2.99,<3.0)",
    "sounddevice (>=0.5.2,<0.6.0)",
    "scipy (>=1.16.1,<2.0.0)",
    "matplotlib (>=3.10.6,<4.0.0)",
    "reportlab (>=4.4.3,<5.0.0)",
    "google-genai (>=1.32.0,<2.0.0)",
    "librosa (>=0.11.0,<0.12.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "livekit-agents[google] (==1.2.0)",
    "livekit-api (>=1.0.2,<2.0.0)",
    "google-ai-generativelanguage (>=0.6.10,<1.0.0)",
    "quart-jwt-extended (>=0.1.0,<0.2.0)",
    "bcrypt (>=4.3.0,<5.0.0)"
]

[project.scripts]
navigai = "navigai_api.main:main"

[tool.poetry]
name = "navigai-api"
version = "0.1.0"
description = "AI-powered career navigation and mock interview platform"
authors = ["Shashidhar Kundur <shashidharkundur1@gmail.com>"]
readme = "README.md"
packages = [{include = "navigai_api", from = "src"}]

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pre-commit = "^4.3.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']